apps/gateway/requirements.txt
# VeloZ Gateway Python Dependencies

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# PostgreSQL support (optional, for production)
psycopg2-binary>=2.9.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Type checking
mypy>=1.0.0
types-requests>=2.28.0
